requirements.txt
fastapi
uvicorn
sqlalchemy
Pillow
logging
